Change button color in serviceinfo form and upload pdf, rename component folder
